# Infofin Solution/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Infofin Solutions LLP API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Consultation {
  id: number;
  name: string;
  /** @nullable */
  companyName?: string | null;
  email: string;
  phone: string;
  serviceInterested: string;
  /** @nullable */
  message?: string | null;
  status: string;
  createdAt: string;
}

export interface ConsultationInput {
  name: string;
  companyName?: string;
  email: string;
  phone: string;
  serviceInterested: string;
  message?: string;
}

export interface ConsultationUpdate {
  status?: string;
  message?: string;
}

export interface Contact {
  id: number;
  name: string;
  /** @nullable */
  companyName?: string | null;
  email: string;
  phone: string;
  subject: string;
  message: string;
  status: string;
  createdAt: string;
}

export interface ContactInput {
  name: string;
  companyName?: string;
  email: string;
  phone: string;
  subject: string;
  message: string;
}

export interface ContactUpdate {
  status?: string;
}

export interface BlogPost {
  id: number;
  title: string;
  slug: string;
  excerpt: string;
  content: string;
  category: string;
  author: string;
  /** @nullable */
  coverImage?: string | null;
  published: boolean;
  createdAt: string;
}

export interface BlogPostInput {
  title: string;
  slug: string;
  excerpt: string;
  content: string;
  category: string;
  author: string;
  coverImage?: string;
  published?: boolean;
}

export interface BlogPostUpdate {
  title?: string;
  slug?: string;
  excerpt?: string;
  content?: string;
  category?: string;
  author?: string;
  coverImage?: string;
  published?: boolean;
}

export interface Career {
  id: number;
  title: string;
  department: string;
  location: string;
  type: string;
  description: string;
  /** @nullable */
  requirements?: string | null;
  active: boolean;
  createdAt: string;
}

export interface CareerInput {
  title: string;
  department: string;
  location: string;
  type: string;
  description: string;
  requirements?: string;
  active?: boolean;
}

export interface CareerUpdate {
  title?: string;
  department?: string;
  location?: string;
  type?: string;
  description?: string;
  requirements?: string;
  active?: boolean;
}

export interface Testimonial {
  id: number;
  clientName: string;
  company: string;
  designation: string;
  content: string;
  rating: number;
  /** @nullable */
  avatar?: string | null;
  active: boolean;
  createdAt: string;
}

export interface TestimonialInput {
  clientName: string;
  company: string;
  designation: string;
  content: string;
  rating: number;
  avatar?: string;
  active?: boolean;
}

export interface TestimonialUpdate {
  clientName?: string;
  company?: string;
  designation?: string;
  content?: string;
  rating?: number;
  avatar?: string;
  active?: boolean;
}

export interface Service {
  id: number;
  name: string;
  category: string;
  description: string;
  /** @nullable */
  icon?: string | null;
  /** @nullable */
  features?: string | null;
  active: boolean;
  createdAt: string;
}

export interface ServiceInput {
  name: string;
  category: string;
  description: string;
  icon?: string;
  features?: string;
  active?: boolean;
}

export interface ServiceUpdate {
  name?: string;
  category?: string;
  description?: string;
  icon?: string;
  features?: string;
  active?: boolean;
}

export interface AdminCredentials {
  username: string;
  password: string;
}

export interface AdminUser {
  id: number;
  username: string;
  role: string;
}

export interface AdminSession {
  success: boolean;
  user: AdminUser;
}

export interface AdminStats {
  totalConsultations: number;
  totalContacts: number;
  newLeads: number;
  pendingConsultations: number;
  totalBlogPosts: number;
  totalTestimonials: number;
  activeCareers: number;
  recentConsultations: Consultation[];
  recentContacts: Contact[];
}