# Infofin Solution/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Infofin Solutions LLP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /consultations:
    get:
      operationId: listConsultations
      tags: [consultations]
      summary: List all consultation bookings
      responses:
        "200":
          description: List of consultations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Consultation"
    post:
      operationId: createConsultation
      tags: [consultations]
      summary: Book a consultation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsultationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Consultation"
  /consultations/{id}:
    get:
      operationId: getConsultation
      tags: [consultations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Consultation detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Consultation"
        "404":
          description: Not found
    patch:
      operationId: updateConsultation
      tags: [consultations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsultationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Consultation"
    delete:
      operationId: deleteConsultation
      tags: [consultations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Contacts
  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List all contact inquiries
      responses:
        "200":
          description: List of contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Submit a contact inquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /contacts/{id}:
    patch:
      operationId: updateContact
      tags: [contacts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    delete:
      operationId: deleteContact
      tags: [contacts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Blog Posts
  /blog-posts:
    get:
      operationId: listBlogPosts
      tags: [blog]
      summary: List blog posts
      responses:
        "200":
          description: List of blog posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlogPost"
    post:
      operationId: createBlogPost
      tags: [blog]
      summary: Create a blog post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogPostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
  /blog-posts/{id}:
    get:
      operationId: getBlogPost
      tags: [blog]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Blog post detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
        "404":
          description: Not found
    patch:
      operationId: updateBlogPost
      tags: [blog]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogPostUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
    delete:
      operationId: deleteBlogPost
      tags: [blog]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Careers
  /careers:
    get:
      operationId: listCareers
      tags: [careers]
      summary: List career listings
      responses:
        "200":
          description: List of careers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Career"
    post:
      operationId: createCareer
      tags: [careers]
      summary: Create a career listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CareerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Career"
  /careers/{id}:
    patch:
      operationId: updateCareer
      tags: [careers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CareerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Career"
    delete:
      operationId: deleteCareer
      tags: [careers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Testimonials
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCredentials"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Invalid credentials
...
  /admin/me:
    get:
      operationId: getAdminMe
      tags: [admin]
      summary: Get current admin user
      responses:
        "200":
          description: Admin user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Not authenticated
...
components:
  schemas:
...
    Consultation:
      type: object
      required: [id, name, email, phone, serviceInterested, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        companyName:
          type: ["string", "null"]
        email:
          type: string
        phone:
          type: string
        serviceInterested:
          type: string
        message:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
    ConsultationInput:
      type: object
      required: [name, email, phone, serviceInterested]
      properties:
        name:
          type: string
        companyName:
          type: string
        email:
          type: string
        phone:
          type: string
        serviceInterested:
          type: string
        message:
          type: string
    ConsultationUpdate:
      type: object
      properties:
        status:
          type: string
        message:
          type: string

    # Contact
    Contact:
      type: object
      required: [id, name, email, phone, subject, message, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        companyName:
          type: ["string", "null"]
        email:
          type: string
        phone:
          type: string
        subject:
          type: string
        message:
          type: string
        status:
          type: string
        createdAt:
          type: string
    ContactInput:
      type: object
      required: [name, email, phone, subject, message]
      properties:
        name:
          type: string
        companyName:
          type: string
        email:
          type: string
        phone:
          type: string
        subject:
          type: string
        message:
          type: string
    ContactUpdate:
      type: object
      properties:
        status:
          type: string

    # Blog Post
    BlogPost:
      type: object
      required: [id, title, slug, excerpt, content, category, author, published, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        content:
          type: string
        category:
          type: string
        author:
          type: string
        coverImage:
          type: ["string", "null"]
        published:
          type: boolean
        createdAt:
          type: string
    BlogPostInput:
      type: object
      required: [title, slug, excerpt, content, category, author]
      properties:
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        content:
          type: string
        category:
          type: string
        author:
          type: string
        coverImage:
          type: string
        published:
          type: boolean
    BlogPostUpdate:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        content:
          type: string
        category:
          type: string
        author:
          type: string
        coverImage:
          type: string
        published:
          type: boolean

    # Career
    Career:
      type: object
      required: [id, title, department, location, type, description, active, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        department:
          type: string
        location:
          type: string
        type:
          type: string
        description:
          type: string
        requirements:
          type: ["string", "null"]
        active:
          type: boolean
        createdAt:
          type: string
    CareerInput:
      type: object
      required: [title, department, location, type, description]
      properties:
        title:
          type: string
        department:
          type: string
        location:
          type: string
        type:
          type: string
        description:
          type: string
        requirements:
          type: string
        active:
          type: boolean
    CareerUpdate:
      type: object
      properties:
        title:
          type: string
        department:
          type: string
        location:
          type: string
        type:
          type: string
        description:
          type: string
        requirements:
          type: string
        active:
          type: boolean

    # Testimonial
...
    AdminCredentials:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AdminSession:
      type: object
      required: [success, user]
      properties:
        success:
          type: boolean
        user:
          $ref: "#/components/schemas/AdminUser"
    AdminUser:
      type: object
      required: [id, username, role]
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
...