# Infofin Solution/lib/api-zod/src/generated/types/blogPost.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Infofin Solutions LLP API
 * OpenAPI spec version: 0.1.0
 */

export interface BlogPost {
  id: number;
  title: string;
  slug: string;
  excerpt: string;
  content: string;
  category: string;
  author: string;
  /** @nullable */
  coverImage?: string | null;
  published: boolean;
  createdAt: string;
}