# Infofin Solution/lib/api-zod/src/generated/types/blogPostUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Infofin Solutions LLP API
 * OpenAPI spec version: 0.1.0
 */

export interface BlogPostUpdate {
  title?: string;
  slug?: string;
  excerpt?: string;
  content?: string;
  category?: string;
  author?: string;
  coverImage?: string;
  published?: boolean;
}