# Infofin Solution/lib/api-zod/src/generated/types/careerUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Infofin Solutions LLP API
 * OpenAPI spec version: 0.1.0
 */

export interface CareerUpdate {
  title?: string;
  department?: string;
  location?: string;
  type?: string;
  description?: string;
  requirements?: string;
  active?: boolean;
}